+++ contracts/library/TimeManagement.sol
@@ -10,8 +10,8 @@ library TimeManagment{
         return _seconds;
     }
 
-    function checkDatalistAndExpired(uint[] memory dataList, uint expireData) internal view returns (bool){
-        uint actualData = block.timestamp;
+    function checkDatalistAndExpired(uint64[] memory dataList, uint64 expireData) internal view returns (bool){
+        uint64 actualData = uint64(block.timestamp);
         for(uint i=1; i < dataList.length;i++){
 
             if(dataList[i]<= actualData){
